CB-365 feat: add location permission handler
위치권한 핸들러 추가
웹뷰 클래스 파일 분리
